+++ src/roq/binance_futures/gateway/order_entry_portfolio.hpp
@@ -250,6 +250,8 @@ struct OrderEntryPortfolio final : public OrderEntry, public web::rest::Client::
   bool download_trades_ = false;
   std::string encode_buffer_;
   bool download_trades_is_first_ = true;
+  //
+  std::string external_order_id_;
 };
 
 }  // namespace gateway