+++ src/roq/binance_futures/gateway/rest_trade.hpp
@@ -164,6 +164,8 @@ struct RestTrade final : public web::rest::Client::Handler {
   bool download_trades_ = false;
   std::string encode_buffer_;
   bool download_trades_is_first_ = true;
+  //
+  std::string external_order_id_;
 };
 
 }  // namespace gateway