+++ src/roq/binance_futures/protocol/json/utils.cpp
@@ -39,6 +39,14 @@ TradingStatus trading_status_helper(SymbolStatus symbol_status, ContractStatus c
   return map(contract_status);
 }
 
+std::string_view get_order_id(std::string_view const &external_order_id) {
+  auto pos = external_order_id.find('#');
+  if (pos == std::string_view::npos) {
+    return external_order_id;
+  }
+  return external_order_id.substr(pos + 1);
+}
+
 }  // namespace json
 }  // namespace protocol
 }  // namespace binance_futures