+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes will be documented in this file.
 
 ### Fixed
 
+* Exchange order ids are only unique by symbol (#608)
 * Incorrect error/text from rejected modify requests (#605)
 * Incorrect OrderAck from modify response when order was canceled (#603)